COMP: Fix cmake option for CUDACOMMON_CUDA_VERSION
CMake "option" only allows boolean ON|OFF values to
be set as cmake CACHE variables.  The
value of CUDACOMMON_CUDA_VERSION is expected to be
a VERSION string like 12.9 or 11.2 which is not
compatible with the cmake "option".

The subsequent
find_package(CUDAToolkit EXACT ${CUDACOMMON_CUDA_VERSION})

Requires the value after EXACT to be a valid version
string (i.e. ON or OFF are invalid choices).

If not set from the command line, default the value
of CUDACOMMON_CUDA_VERSION to the version that was
previously found.